+++ plugin/src/main/java/com/stratio/cassandra/lucene/schema/mapping/LongMapper.java
@@ -23,6 +23,7 @@ import org.apache.lucene.search.SortField;
 import org.apache.lucene.search.SortField.Type;
 import org.apache.lucene.search.SortedNumericSortField;
 
+import java.util.Date;
 import java.util.Optional;
 
 /**
@@ -47,7 +48,7 @@ public class LongMapper extends SingleColumnMapper.SingleFieldMapper<Long> {
      * @param boost the boost
      */
     public LongMapper(String field, String column, Boolean validated, Float boost) {
-        super(field, column, true, validated, null, Long.class, NUMERIC_TYPES);
+        super(field, column, true, validated, null, Long.class, NUMERIC_TYPES_WITH_DATE);
         this.boost = boost == null ? DEFAULT_BOOST : boost;
     }
 
@@ -56,7 +57,9 @@ public class LongMapper extends SingleColumnMapper.SingleFieldMapper<Long> {
     protected Long doBase(String name, Object value) {
         if (value instanceof Number) {
             return ((Number) value).longValue();
-        } else if (value instanceof String) {
+        } else if (value instanceof Date) {
+            return ((Date) value).getTime();
+        } if (value instanceof String) {
             try {
                 return Double.valueOf((String) value).longValue();
             } catch (NumberFormatException e) {