+++ plugin/src/test/java/com/stratio/cassandra/lucene/schema/mapping/LongMapperTest.java
@@ -22,6 +22,9 @@ import org.apache.lucene.index.DocValuesType;
 import org.apache.lucene.search.SortField;
 import org.junit.Test;
 
+import java.util.Date;
+
+import static com.stratio.cassandra.lucene.schema.SchemaBuilders.integerMapper;
 import static com.stratio.cassandra.lucene.schema.SchemaBuilders.longMapper;
 import static org.junit.Assert.*;
 
@@ -128,7 +131,6 @@ public class LongMapperTest extends AbstractMapperTest {
         LongMapper mapper = longMapper().boost(1f).build("field");
         Long parsed = mapper.base("test", 3.5f);
         assertEquals("Base for float is wrong", Long.valueOf(3), parsed);
-
     }
 
     @Test
@@ -136,7 +138,6 @@ public class LongMapperTest extends AbstractMapperTest {
         LongMapper mapper = longMapper().boost(1f).build("field");
         Long parsed = mapper.base("test", 3.6f);
         assertEquals("Base for float is wrong", Long.valueOf(3), parsed);
-
     }
 
     @Test
@@ -159,7 +160,6 @@ public class LongMapperTest extends AbstractMapperTest {
         LongMapper mapper = longMapper().boost(1f).build("field");
         Long parsed = mapper.base("test", 3.6d);
         assertEquals("Base for double is wrong", Long.valueOf(3), parsed);
-
     }
 
     @Test
@@ -174,7 +174,6 @@ public class LongMapperTest extends AbstractMapperTest {
         LongMapper mapper = longMapper().boost(1f).build("field");
         Long parsed = mapper.base("test", "3.2");
         assertEquals("Base for string is wrong", Long.valueOf(3), parsed);
-
     }
 
     @Test
@@ -182,6 +181,13 @@ public class LongMapperTest extends AbstractMapperTest {
         LongMapper mapper = longMapper().boost(1f).build("field");
         Long parsed = mapper.base("test", "3.2");
         assertEquals("Base for string is wrong", Long.valueOf(3), parsed);
+    }
+
+    @Test
+    public void testValueWithDate() {
+        LongMapper mapper = longMapper().boost(1f).build("field");
+        Long parsed = mapper.base("test", new Date(10));
+        assertEquals("Base for dates is wrong", Long.valueOf(10), parsed);
     }
 
     @Test